# First Branch Commit #
This branch is meant for adding a new feature to the framework: Elements
It will allow for greater control over relationships between attacker
and defender, as well as make for dynamic balancing possibilities

Renamed Global_Elements to Global_Config
Created elements and damage_types dicts in Global_Config
Created modifier_dict constructors for elements and damage_types
